+++ LibAcousticSensing/Matlab/Functions/CreateStandaloneContextAndUser.m
@@ -5,17 +5,17 @@ function [context, user] = CreateStandaloneContextAndUser()
 % TODO: find a smarter way to get these values directly from the SensingServer class    
 
 % define the context for standalone mode
-context.mode = int16(SensingServer.MODE_STANDALONE);
+context.mode = int32(SensingServer.MODE_STANDALONE);
 
 % Types of user-defined callback messages 
-context.CALLBACK_TYPE_ERROR = int16(SensingServer.CALLBACK_TYPE_ERROR);
-context.CALLBACK_TYPE_INIT  = int16(SensingServer.CALLBACK_TYPE_INIT);
-context.CALLBACK_TYPE_DATA  = int16(SensingServer.CALLBACK_TYPE_DATA);
-context.CALLBACK_TYPE_USER  = int16(SensingServer.CALLBACK_TYPE_USER);
+context.CALLBACK_TYPE_ERROR = int32(SensingServer.CALLBACK_TYPE_ERROR);
+context.CALLBACK_TYPE_INIT  = int32(SensingServer.CALLBACK_TYPE_INIT);
+context.CALLBACK_TYPE_DATA  = int32(SensingServer.CALLBACK_TYPE_DATA);
+context.CALLBACK_TYPE_USER  = int32(SensingServer.CALLBACK_TYPE_USER);
 
 % Here we define the user data
-user.code = int16(5566);
-user.valInt = int16(5566);
+user.code = int32(5566);
+user.valInt = int32(5566);
 
 
 end
